+++ zerver/tests/test_events.py
@@ -3050,6 +3050,29 @@ class NormalActionsTest(BaseAction):
 
     def test_do_deactivate_user(self) -> None:
         user_profile = self.example_user("cordelia")
+        members_group = NamedUserGroup.objects.get(
+            name=SystemGroups.MEMBERS, realm=user_profile.realm, is_system_group=True
+        )
+        setting_group = self.create_or_update_anonymous_group_for_setting(
+            [user_profile], [members_group]
+        )
+        do_change_realm_permission_group_setting(
+            self.user_profile.realm,
+            "can_create_public_channel_group",
+            setting_group,
+            acting_user=None,
+        )
+        hamletcharacters_group = NamedUserGroup.objects.get(
+            name="hamletcharacters", realm=self.user_profile.realm
+        )
+        hamlet = self.example_user("hamlet")
+        setting_group = self.create_or_update_anonymous_group_for_setting(
+            [user_profile, hamlet], [members_group]
+        )
+        do_change_user_group_permission_setting(
+            hamletcharacters_group, "can_mention_group", setting_group, acting_user=None
+        )
+
         with self.verify_action(num_events=1) as events:
             do_deactivate_user(user_profile, acting_user=None)
         check_realm_user_update("events[0]", events[0], "is_active")
@@ -3070,19 +3093,53 @@ class NormalActionsTest(BaseAction):
         # event if they cannot access the deactivated user.
         user_profile = self.example_user("cordelia")
         self.user_profile = self.example_user("polonius")
-        with self.verify_action(num_events=3) as events:
+        with self.verify_action(num_events=6) as events:
             do_deactivate_user(user_profile, acting_user=None)
         check_user_group_remove_members("events[0]", events[0])
         check_user_group_remove_members("events[1]", events[1])
         check_user_group_remove_members("events[2]", events[2])
+        check_user_group_update("events[3]", events[3], "can_manage_group")
+        check_realm_update_dict("events[4]", events[4])
+        check_user_group_update("events[5]", events[5], "can_mention_group")
+        self.assertEqual(
+            events[3]["data"]["can_manage_group"],
+            AnonymousSettingGroupDict(direct_members=[], direct_subgroups=[]),
+        )
+        self.assertEqual(
+            events[4]["data"]["can_create_public_channel_group"],
+            AnonymousSettingGroupDict(direct_members=[], direct_subgroups=[members_group.id]),
+        )
+        self.assertEqual(
+            events[5]["data"]["can_mention_group"],
+            AnonymousSettingGroupDict(
+                direct_members=[hamlet.id], direct_subgroups=[members_group.id]
+            ),
+        )
 
         user_profile = self.example_user("cordelia")
         do_reactivate_user(user_profile, acting_user=None)
-        with self.verify_action(num_events=3, user_list_incomplete=True) as events:
+        with self.verify_action(num_events=6, user_list_incomplete=True) as events:
             do_deactivate_user(user_profile, acting_user=None)
         check_user_group_remove_members("events[0]", events[0])
         check_user_group_remove_members("events[1]", events[1])
         check_user_group_remove_members("events[2]", events[2])
+        check_user_group_update("events[3]", events[3], "can_manage_group")
+        check_realm_update_dict("events[4]", events[4])
+        check_user_group_update("events[5]", events[5], "can_mention_group")
+        self.assertEqual(
+            events[3]["data"]["can_manage_group"],
+            AnonymousSettingGroupDict(direct_members=[], direct_subgroups=[]),
+        )
+        self.assertEqual(
+            events[4]["data"]["can_create_public_channel_group"],
+            AnonymousSettingGroupDict(direct_members=[], direct_subgroups=[members_group.id]),
+        )
+        self.assertEqual(
+            events[5]["data"]["can_mention_group"],
+            AnonymousSettingGroupDict(
+                direct_members=[hamlet.id], direct_subgroups=[members_group.id]
+            ),
+        )
 
         user_profile = self.example_user("shiva")
         with self.verify_action(num_events=1) as events:
@@ -3092,12 +3149,17 @@ class NormalActionsTest(BaseAction):
         # Guest loses access to deactivated user if the user
         # was not involved in DMs.
         user_profile = self.example_user("hamlet")
-        with self.verify_action(num_events=4) as events:
+        with self.verify_action(num_events=5) as events:
             do_deactivate_user(user_profile, acting_user=None)
         check_user_group_remove_members("events[0]", events[0])
         check_user_group_remove_members("events[1]", events[1])
         check_user_group_remove_members("events[2]", events[2])
-        check_realm_user_remove("events[3]]", events[3])
+        check_user_group_update("events[3]", events[3], "can_mention_group")
+        check_realm_user_remove("events[4]]", events[4])
+        self.assertEqual(
+            events[3]["data"]["can_mention_group"],
+            AnonymousSettingGroupDict(direct_members=[], direct_subgroups=[members_group.id]),
+        )
 
         user_profile = self.example_user("aaron")
         # One update event is for a deactivating a bot owned by aaron.
@@ -3143,6 +3205,30 @@ class NormalActionsTest(BaseAction):
         check_subscription_peer_remove("events[4]", events[4])
         check_stream_delete("events[5]", events[5])
 
+        user_profile = self.example_user("cordelia")
+        members_group = NamedUserGroup.objects.get(
+            name=SystemGroups.MEMBERS, realm=user_profile.realm, is_system_group=True
+        )
+        hamletcharacters_group = NamedUserGroup.objects.get(
+            name="hamletcharacters", realm=self.user_profile.realm
+        )
+
+        setting_group = self.create_or_update_anonymous_group_for_setting(
+            [user_profile], [hamletcharacters_group]
+        )
+        do_change_realm_permission_group_setting(
+            user_profile.realm,
+            "can_create_public_channel_group",
+            setting_group,
+            acting_user=None,
+        )
+        setting_group = self.create_or_update_anonymous_group_for_setting(
+            [user_profile], [members_group]
+        )
+        do_change_user_group_permission_setting(
+            hamletcharacters_group, "can_mention_group", setting_group, acting_user=None
+        )
+
         self.set_up_db_for_testing_user_access()
         # Test that guest users receive realm_user/update event
         # only if they can access the reactivated user.
@@ -3152,11 +3238,30 @@ class NormalActionsTest(BaseAction):
         self.user_profile = self.example_user("polonius")
         # Guest users receives group members update event for three groups -
         # members group, full members group and hamletcharacters group.
-        with self.verify_action(num_events=3) as events:
+        with self.verify_action(num_events=6) as events:
             do_reactivate_user(user_profile, acting_user=None)
         check_user_group_add_members("events[0]", events[0])
         check_user_group_add_members("events[1]", events[1])
         check_user_group_add_members("events[2]", events[2])
+        check_user_group_update("events[3]", events[3], "can_manage_group")
+        check_realm_update_dict("events[4]", events[4])
+        check_user_group_update("events[5]", events[5], "can_mention_group")
+        self.assertEqual(
+            events[3]["data"]["can_manage_group"],
+            AnonymousSettingGroupDict(direct_members=[user_profile.id], direct_subgroups=[]),
+        )
+        self.assertEqual(
+            events[4]["data"]["can_create_public_channel_group"],
+            AnonymousSettingGroupDict(
+                direct_members=[user_profile.id], direct_subgroups=[hamletcharacters_group.id]
+            ),
+        )
+        self.assertEqual(
+            events[5]["data"]["can_mention_group"],
+            AnonymousSettingGroupDict(
+                direct_members=[user_profile.id], direct_subgroups=[members_group.id]
+            ),
+        )
 
         user_profile = self.example_user("shiva")
         do_deactivate_user(user_profile, acting_user=None)